Remove the "Hive" hashtag tab on Blog.D.Buzz and add "Events" and "FOR SALE"The way it works is that the tabs at the top of deskop view in Blog.D.Buzz are basically just hashtag buttons.
The buttons are meant to act exactly like it would act if the the words were hashtags that were clicked on.

Keep in mind, hashtags should always be rendered at as all non-lower case at the blockchain level to prevent dividing hashtag feeds in two based on capitalization.

Anyhow, kindly remove the “Hive” tab and then add “Events” and “For Sale” (#ForSale). This means the top tab menu will read as follows.

Trending Following Latest News Events FOR SALE